removed translator addOn specific stuff and replaced it with a more consistent model
And the model is: Every translation is done via sly_I18N. In the backend,
everything stays the same. For the frontend, we load a base database based on
the default backend locale. This DB is only used for error messages during
boot-up. The *real* language files live in develop/lang and are named after their
correspondig clang locale.
AddOns like the translator now serve as more of a nice backend for those
first-class languages files in the frontend. They probably have to be slightly
rewritten, though.