fixed handling of non-existing articles in sly_Util_Article::getUrl()
Before this patch, if you use getUrl() on a non-existing ID, the getArticle()
would return the ID and not null, because the ID is also given as the default
value (to enable calls like getUrl(CURRENT_ARTICLE)).
The check for "not article" is therefore broken, as it doesn't catch $article
being a non-null integer.

The resolveConstant was added accidentally. But I left it, maybe it becomes
useful one day.